+++ server/src/routes/user/user.service.ts
@@ -1,15 +1,15 @@
 import { Injectable } from '@nestjs/common';
 import { UserRepository } from '../../database/models/user/user.repository';
-import { ImageService } from '../../services/image/image.service';
-import { IImage } from '../../services/image/interfaces/IImage';
+import { FileService } from '../../services/file/image.service';
+import { IImage } from '../../services/file/interfaces/IImage';
 import { Request } from 'express';
 
 @Injectable()
 export class UserService {
-    constructor(private readonly _imageService: ImageService, private readonly _userRepository: UserRepository) {}
+    constructor(private readonly _fileService: FileService, private readonly _userRepository: UserRepository) {}
 
     public async updateAvatar(request: Request, image: IImage): Promise<void> {
-        const imageName = await this._imageService.uploadImage(image);
+        const imageName = await this._fileService.uploadImage(image);
         await this._changeAvatarInDatabase(request.user.id, imageName);
     }
