+++ server/src/services/file/interfaces/ICloudProvider.ts
