+++ server/src/services/file/interfaces/IImage.ts
